+++ samples/SATokenSample.py
@@ -1,9 +1,9 @@
 from skyflow.Errors import SkyflowError
-from skyflow.ServiceAccount import GenerateToken
+from skyflow.ServiceAccount import GenerateBearerToken
 
 
 try:
-    accessToken, tokenType = GenerateToken('<YOUR_CREDENTIALS_FILE_PATH>')
+    accessToken, tokenType = GenerateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
     print("Access Token:", accessToken)
     print("Type of token:", tokenType)
 except SkyflowError as e: