+++ skyflow/Errors/_skyflowErrors.py
@@ -1,4 +1,5 @@
 from enum import Enum
+from skyflow._utils import log_error
 
 class SkyflowErrorCodes(Enum):
     INVALID_INPUT = 400
@@ -8,6 +9,7 @@ class SkyflowErrorCodes(Enum):
 class SkyflowErrorMessages(Enum):
     FILE_NOT_FOUND = "File at %s not found"
     FILE_INVALID_JSON = "File at %s is not in JSON format"
+    INVALID_CREDENTIALS = "Given credentials are not valid"
     INVALID_URL = "Given url '%s' is invalid"
 
     MISSING_PRIVATE_KEY = "Unable to read Private key"
@@ -50,8 +52,11 @@ class SkyflowErrorMessages(Enum):
     VAULT_URL_INVALID_TYPE = "Expected Vault URL to be str, got %s"
     TOKEN_PROVIDER_ERROR = "Expected Token Provider to be function, got %s"
 
+    EMPTY_VAULT_ID = "Vault ID must not be empty"
+    EMPTY_VAULT_URL = "Vault URL must not be empty"
+
 class SkyflowError(Exception):
-    def __init__(self, code, message="An Error occured", data={}) -> None:
+    def __init__(self, code, message="An Error occured", data={}, interface: str=None) -> None:
         if type(code) is SkyflowErrorCodes:
             self.code = code.value
         else:
@@ -60,6 +65,7 @@ class SkyflowError(Exception):
             self.message = message.value
         else:
             self.message = message
+        log_error(self.message, interface)
         self.data = data
         super().__init__(self.message)