+++ skyflow/ServiceAccount/_token.py
@@ -3,13 +3,23 @@ import json
 import jwt
 import datetime
 import requests
-from requests import status_codes
+from warnings import warn
+from skyflow._utils import log_info, InterfaceName, InfoMessages
+
+from requests.models import Response
 
 from skyflow.Errors._skyflowErrors import *
 
 ResponseToken = namedtuple('ResponseToken', ['AccessToken', 'TokenType'])
 
 def GenerateToken(credentialsFilePath: str) -> ResponseToken:
+    '''
+    This function has been deprecated and replaced with GenerateBearerToken(credentialsFilePath: str)
+    '''
+    warn('This function has been deprecated and replaced with GenerateBearerToken(credentialsFilePath: str)', DeprecationWarning)
+    GenerateBearerToken(credentialsFilePath)
+
+def GenerateBearerToken(credentialsFilePath: str) -> ResponseToken:
 
     '''
     This function is used to get the access token for skyflow Service Accounts.
@@ -19,6 +29,10 @@ def GenerateToken(credentialsFilePath: str) -> ResponseToken:
         1. AccessToken: The access token
         2. TokenType: The type of access token (eg: Bearer)
     '''
+
+    interface = InterfaceName.GENERATE_BEARER_TOKEN.value
+
+    log_info(InfoMessages.GENERATE_BEARER_TOKEN_TRIGGERED.value, interface)
 
     try:
         credentialsFile = open(credentialsFilePath, 'r')
@@ -32,8 +46,35 @@ def GenerateToken(credentialsFilePath: str) -> ResponseToken:
     finally:
         credentialsFile.close()
 
+    result = getSAToken(credentials)
 
-    return getSAToken(credentials)
+    log_info(InfoMessages.GENERATE_BEARER_TOKEN_SUCCESS.value, interface)
+    return result
+
+def GenerateBearerTokenFromCreds(credentials: str) -> ResponseToken:
+
+    '''
+    This function is used to get the access token for skyflow Service Accounts.
+    `credentials` arg takes the content of the credentials file that is downloaded after Service Account creation.
+
+    Response Token is a named tupe with two attributes:
+        1. AccessToken: The access token
+        2. TokenType: The type of access token (eg: Bearer)
+    '''
+
+    interface = InterfaceName.GENERATE_BEARER_TOKEN.value
+
+    log_info(InfoMessages.GENERATE_BEARER_TOKEN_TRIGGERED.value, interface)
+    try:
+        jsonCredentials = json.load(credentials)
+    except Exception as e:
+        raise SkyflowError(SkyflowErrorCodes.INVALID_INPUT, SkyflowErrorMessages.INVALID_CREDENTIALS, interface=interface)
+    result = getSAToken(jsonCredentials)
+
+    log_info(InfoMessages.GENERATE_BEARER_TOKEN_SUCCESS.value, interface=interface)
+    return result
+
+    
 
 def getSAToken(credentials):
     try: