+++ tests/ServiceAccount/testGenerateToken.py
@@ -1,10 +1,10 @@
 import unittest
 import os
 from dotenv import dotenv_values
-from skyflow.ServiceAccount import GenerateToken
+from skyflow.ServiceAccount import GenerateBearerToken
 from skyflow.Errors._skyflowErrors import *
 
-class TestGenerateToken(unittest.TestCase):
+class TestGenerateBearerToken(unittest.TestCase):
 
     def setUp(self) -> None:
         self.dataPath = os.path.join(os.getcwd(), 'tests/ServiceAccount/data/')
@@ -15,7 +15,7 @@ class TestGenerateToken(unittest.TestCase):
 
     def testWithInvalidFilePath(self):
         try:
-            GenerateToken('unknownfilepath')
+            GenerateBearerToken('unknownfilepath')
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.FILE_NOT_FOUND.value % ('unknownfilepath'))
@@ -24,51 +24,51 @@ class TestGenerateToken(unittest.TestCase):
     def testInvalidJSON(self):
         path = self.getDataPath('empty')
         try:
-            GenerateToken(path)
+            GenerateBearerToken(path)
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.FILE_INVALID_JSON.value % (path))
 
     def testWithNoPrivateKey(self):
         try:
-            GenerateToken(self.getDataPath('noPrivateKey'))
+            GenerateBearerToken(self.getDataPath('noPrivateKey'))
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.MISSING_PRIVATE_KEY.value)
     
     def testWithNoClientID(self):
         try:
-            GenerateToken(self.getDataPath('noClientID'))
+            GenerateBearerToken(self.getDataPath('noClientID'))
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.MISSING_CLIENT_ID.value)
 
     def testWithNoKeyID(self):
         try:
-            GenerateToken(self.getDataPath('noKeyID'))
+            GenerateBearerToken(self.getDataPath('noKeyID'))
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.MISSING_KEY_ID.value)
 
     def testWithNoTokenURI(self):
         try:
-            GenerateToken(self.getDataPath('noTokenURI'))
+            GenerateBearerToken(self.getDataPath('noTokenURI'))
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.MISSING_TOKEN_URI.value)
 
     def testInvalidCreds(self):
         try:
-             GenerateToken(self.getDataPath('invalidPrivateKey'))
+             GenerateBearerToken(self.getDataPath('invalidPrivateKey'))
         except SkyflowError as se:
             self.assertEqual(se.code, SkyflowErrorCodes.INVALID_INPUT.value)
             self.assertEqual(se.message, SkyflowErrorMessages.JWT_INVALID_FORMAT.value)
 
-    def testGenerateTokenSuccess(self):
+    def testGenerateBearerTokenSuccess(self):
         env_values = dotenv_values('.env')
         credentials_path = env_values['CREDENTIALS_FILE_PATH']
         try:
-            token, type = GenerateToken(credentials_path)
+            token, type = GenerateBearerToken(credentials_path)
             self.assertIsNotNone(token)
             self.assertEqual(type, 'Bearer')
         except SkyflowError: