+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2022-01-04
+
+### Added
+- Logging functionality
+- `setLogLevel` function for setting the package-level LogLevel
+- `GenerateBearerTokenFromCreds` function which takes credentials as string
+
+### Changed
+- Renamed and deprecated `GenerateToken` in favor of `GenerateBearerToken`
+- Make `vaultID` and `vaultURL` optional in `Client` constructor
+
+
 ## [1.1.0] - 2021-11-10
 
 ### Added